MBS-8472 Move instrumentation-plugin configuration to root build.gradle.kts
