+++ build.gradle.kts
@@ -1,5 +1,9 @@
 @file:Suppress("UnstableApiUsage")
 
+import com.avito.instrumentation.configuration.InstrumentationPluginConfiguration.GradleInstrumentationPluginConfiguration
+import com.avito.kotlin.dsl.getOptionalStringProperty
+import com.avito.utils.gradle.kubernetesCredentials
+
 buildscript {
     buildscript {
         dependencies {
@@ -11,6 +15,7 @@ buildscript {
 plugins {
     id("org.jetbrains.kotlin.jvm") apply false
     id("com.android.application") apply false
+    id("com.avito.android.instrumentation-tests") apply false
 }
 
 /**
@@ -86,6 +91,165 @@ subprojects {
                 isAbortOnError = false
                 isWarningsAsErrors = true
                 textReport = true
+            }
+        }
+    }
+
+    plugins.withType<com.avito.instrumentation.InstrumentationTestsPlugin>() {
+        extensions.getByType<GradleInstrumentationPluginConfiguration>().apply {
+
+            //todo make these params optional features in plugin
+            reportApiUrl = project.getOptionalStringProperty("avito.report.url") ?: "http://stub"
+            reportApiFallbackUrl = project.getOptionalStringProperty("avito.report.fallbackUrl") ?: "http://stub"
+            reportViewerUrl = project.getOptionalStringProperty("avito.report.viewerUrl") ?: "http://stub"
+            registry = project.getOptionalStringProperty("avito.registry", "registry") ?: "registry"
+            sentryDsn =
+                project.getOptionalStringProperty("avito.instrumentaion.sentry.dsn") ?: "http://stub-project@stub-host/0"
+            slackToken = project.getOptionalStringProperty("avito.slack.token") ?: "stub"
+            fileStorageUrl = project.getOptionalStringProperty("avito.fileStorage.url") ?: "http://stub"
+
+            logcatTags = setOf(
+                "UITestRunner:*",
+                "ActivityManager:*",
+                "ReportTestListener:*",
+                "StorageJsonTransport:*",
+                "TestReport:*",
+                "VideoCaptureListener:*",
+                "TestRunner:*",
+                "SystemDialogsManager:*",
+                "AndroidJUnitRunner:*",
+                "ito.android.de:*", //по этому тэгу система пишет логи об использовании hidden/restricted api https://developer.android.com/distribute/best-practices/develop/restrictions-non-sdk-interfaces
+                "*:E"
+            )
+
+            instrumentationParams = mapOf(
+                "videoRecording" to "failed",
+                "jobSlug" to "FunctionalTests"
+            )
+
+
+            val runAllFilterName = "runAll"
+            filters.register(runAllFilterName)
+
+            filters.register("dynamicFilter") {
+                val includeAnnotation: String? = project.getOptionalStringProperty("includeAnnotation", true)
+                if (includeAnnotation != null) {
+                    fromSource.includeByAnnotations(setOf(includeAnnotation))
+                }
+                val includePrefix: String? = project.getOptionalStringProperty("includePrefix", true)
+                if (includePrefix != null) {
+                    fromSource.includeByPrefixes(setOf(includePrefix))
+                }
+            }
+
+            val defaultFilter = "default"
+            val customFilter: String = project.getOptionalStringProperty("localFilter", defaultFilter)
+
+            configurationsContainer.register("Local") {
+                tryToReRunOnTargetBranch = false
+                reportSkippedTests = true
+                reportFlakyTests = false
+                filter = customFilter
+
+                targetsContainer.register("api27") {
+                    deviceName = "API27"
+
+                    scheduling = com.avito.instrumentation.configuration.target.scheduling.SchedulingConfiguration().apply {
+                        quota = com.avito.instrumentation.configuration.target.scheduling.quota.QuotaConfiguration().apply {
+                            retryCount = 1
+                            minimumSuccessCount = 1
+                        }
+
+                        reservation = com.avito.instrumentation.configuration.target.scheduling.reservation.TestsBasedDevicesReservationConfiguration()
+                            .apply {
+                            device = com.avito.instrumentation.reservation.request.Device.LocalEmulator.device(27)
+                            maximum = 1
+                            minimum = 1
+                            testsPerEmulator = 1
+                        }
+                    }
+                }
+            }
+
+            val credentials = project.kubernetesCredentials
+            if (credentials is com.avito.utils.gradle.KubernetesCredentials.Service || credentials is com.avito.utils.gradle.KubernetesCredentials.Config) {
+                configurationsContainer.register("ui") {
+                    tryToReRunOnTargetBranch = false
+                    reportSkippedTests = true
+                    reportFlakyTests = true
+                    filter = customFilter
+
+                    targetsContainer.register("api22") {
+                        deviceName = "API22"
+
+                        scheduling = com.avito.instrumentation.configuration.target.scheduling.SchedulingConfiguration()
+                            .apply {
+                            quota = com.avito.instrumentation.configuration.target.scheduling.quota.QuotaConfiguration()
+                                .apply {
+                                retryCount = 1
+                                minimumSuccessCount = 1
+                            }
+
+                            reservation = com.avito.instrumentation.configuration.target.scheduling.reservation.TestsBasedDevicesReservationConfiguration()
+                                .apply {
+                                device = com.avito.instrumentation.reservation.request.Device.Emulator.Emulator22
+                                maximum = 50
+                                minimum = 2
+                                testsPerEmulator = 3
+                            }
+                        }
+                    }
+
+                    targetsContainer.register("api27") {
+                        deviceName = "API27"
+
+                        scheduling = com.avito.instrumentation.configuration.target.scheduling.SchedulingConfiguration()
+                            .apply {
+                            quota = com.avito.instrumentation.configuration.target.scheduling.quota.QuotaConfiguration()
+                                .apply {
+                                retryCount = 1
+                                minimumSuccessCount = 1
+                            }
+
+                            reservation = com.avito.instrumentation.configuration.target.scheduling.reservation.TestsBasedDevicesReservationConfiguration()
+                                .apply {
+                                device = com.avito.instrumentation.reservation.request.Device.Emulator.Emulator27
+                                maximum = 50
+                                minimum = 2
+                                testsPerEmulator = 3
+                            }
+                        }
+                    }
+                }
+
+                configurationsContainer.register("uiDebug") {
+                    tryToReRunOnTargetBranch = false
+                    reportSkippedTests = false
+                    reportFlakyTests = false
+                    enableDeviceDebug = true
+                    filter = customFilter
+
+                    targetsContainer.register("api27") {
+                        deviceName = "API27"
+
+                        scheduling = com.avito.instrumentation.configuration.target.scheduling.SchedulingConfiguration()
+                            .apply {
+                            quota = com.avito.instrumentation.configuration.target.scheduling.quota.QuotaConfiguration()
+                                .apply {
+                                retryCount = 1
+                                minimumSuccessCount = 1
+                            }
+
+                            reservation = com.avito.instrumentation.configuration.target.scheduling.reservation.TestsBasedDevicesReservationConfiguration()
+                                .apply {
+                                device = com.avito.instrumentation.reservation.request.Device.Emulator.Emulator27
+                                maximum = 1
+                                minimum = 1
+                                testsPerEmulator = 1
+                            }
+                        }
+                    }
+                }
             }
         }
     }
