+++ samples/test-app-kaspresso/build.gradle.kts
@@ -1,13 +1,6 @@
+
 import com.android.build.gradle.ProguardFiles.ProguardFile
-import com.avito.instrumentation.configuration.InstrumentationPluginConfiguration.GradleInstrumentationPluginConfiguration
-import com.avito.instrumentation.configuration.target.scheduling.SchedulingConfiguration
-import com.avito.instrumentation.configuration.target.scheduling.quota.QuotaConfiguration
-import com.avito.instrumentation.configuration.target.scheduling.reservation.TestsBasedDevicesReservationConfiguration
-import com.avito.instrumentation.reservation.request.Device.Emulator.Emulator22
-import com.avito.instrumentation.reservation.request.Device.Emulator.Emulator27
-import com.avito.instrumentation.reservation.request.Device.LocalEmulator
 import com.avito.kotlin.dsl.getOptionalStringProperty
-import com.avito.utils.gradle.buildEnvironment
 
 plugins {
     id("com.android.application")
@@ -86,144 +79,3 @@ dependencies(delegateClosureOf<DependencyHandler> {
 
     androidTestImplementation(Dependencies.test.junit)
 })
-
-extensions.getByType<GradleInstrumentationPluginConfiguration>().apply {
-
-    //todo make these params optional features in plugin
-    reportApiUrl = project.getOptionalStringProperty("avito.report.url") ?: "http://stub"
-    reportApiFallbackUrl = project.getOptionalStringProperty("avito.report.fallbackUrl") ?: "http://stub"
-    reportViewerUrl = project.getOptionalStringProperty("avito.report.viewerUrl") ?: "http://stub"
-    registry = project.getOptionalStringProperty("avito.registry") ?: "registry"
-    sentryDsn =
-        project.getOptionalStringProperty("avito.instrumentaion.sentry.dsn") ?: "http://stub-project@stub-host/0"
-    slackToken = project.getOptionalStringProperty("avito.slack.token") ?: "stub"
-    fileStorageUrl = project.getOptionalStringProperty("avito.fileStorage.url") ?: "http://stub"
-    // deprecated since 2020.3.6
-    output = project.rootProject.file("outputs/${project.name}/instrumentation").path
-
-    logcatTags = setOf(
-        "UITestRunner:*",
-        "ActivityManager:*",
-        "ReportTestListener:*",
-        "StorageJsonTransport:*",
-        "TestReport:*",
-        "VideoCaptureListener:*",
-        "TestRunner:*",
-        "SystemDialogsManager:*",
-        "AndroidJUnitRunner:*",
-        "ito.android.de:*", //по этому тэгу система пишет логи об использовании hidden/restricted api https://developer.android.com/distribute/best-practices/develop/restrictions-non-sdk-interfaces
-        "*:E"
-    )
-
-    instrumentationParams = mapOf(
-        "videoRecording" to "failed",
-        "jobSlug" to "FunctionalTests"
-    )
-
-    filters.register("all")
-
-    configurationsContainer.register("Local") {
-        tryToReRunOnTargetBranch = false
-        reportSkippedTests = true
-        reportFlakyTests = false
-        filter = "all"
-
-        targetsContainer.register("api27") {
-            deviceName = "API27"
-
-            scheduling = SchedulingConfiguration().apply {
-                quota = QuotaConfiguration().apply {
-                    retryCount = 1
-                    minimumSuccessCount = 1
-                }
-
-                reservation = TestsBasedDevicesReservationConfiguration().apply {
-                    device = LocalEmulator.device(27)
-                    maximum = 1
-                    minimum = 1
-                    testsPerEmulator = 1
-                }
-            }
-        }
-    }
-    // TODO uncomment after 2020.4.6 release
-//    val credentials = project.kubernetesCredentials
-//    if (credentials is KubernetesCredentials.Service || credentials is KubernetesCredentials.Config) {
-//    }
-    if (project.buildEnvironment is com.avito.utils.gradle.BuildEnvironment.CI) {
-        configurationsContainer.register("ui") {
-            tryToReRunOnTargetBranch = false
-            reportSkippedTests = true
-            reportFlakyTests = true
-
-            targetsContainer.register("api22") {
-                deviceName = "API22"
-
-                scheduling = SchedulingConfiguration().apply {
-                    quota = QuotaConfiguration().apply {
-                        retryCount = 1
-                        minimumSuccessCount = 1
-                    }
-
-                    reservation = TestsBasedDevicesReservationConfiguration().apply {
-                        device = Emulator22
-                        maximum = 50
-                        minimum = 2
-                        testsPerEmulator = 3
-                    }
-                }
-            }
-
-            targetsContainer.register("api27") {
-                deviceName = "API27"
-
-                scheduling = SchedulingConfiguration().apply {
-                    quota = QuotaConfiguration().apply {
-                        retryCount = 1
-                        minimumSuccessCount = 1
-                    }
-
-                    reservation = TestsBasedDevicesReservationConfiguration().apply {
-                        device = Emulator27
-                        maximum = 50
-                        minimum = 2
-                        testsPerEmulator = 3
-                    }
-                }
-            }
-        }
-
-        configurationsContainer.register("uiDebug") {
-            tryToReRunOnTargetBranch = false
-            reportSkippedTests = false
-            reportFlakyTests = false
-            enableDeviceDebug = true
-
-            targetsContainer.register("api27") {
-                deviceName = "API27"
-
-                scheduling = SchedulingConfiguration().apply {
-                    quota = QuotaConfiguration().apply {
-                        retryCount = 1
-                        minimumSuccessCount = 1
-                    }
-
-                    reservation = TestsBasedDevicesReservationConfiguration().apply {
-                        device = Emulator27
-                        maximum = 1
-                        minimum = 1
-                        testsPerEmulator = 1
-                    }
-                }
-            }
-        }
-    }
-}
-
-configurations.all {
-    if (name.contains("AndroidTestRuntimeClasspath")) {
-        resolutionStrategy {
-            force("org.jetbrains:annotations:16.0.1")
-        }
-    }
-}
