+++ samples/test-app/build.gradle.kts
@@ -1,14 +1,6 @@
+
 import com.android.build.gradle.ProguardFiles.ProguardFile
-import com.avito.instrumentation.configuration.InstrumentationPluginConfiguration.GradleInstrumentationPluginConfiguration
-import com.avito.instrumentation.configuration.target.scheduling.SchedulingConfiguration
-import com.avito.instrumentation.configuration.target.scheduling.quota.QuotaConfiguration
-import com.avito.instrumentation.configuration.target.scheduling.reservation.TestsBasedDevicesReservationConfiguration
-import com.avito.instrumentation.reservation.request.Device.Emulator.Emulator22
-import com.avito.instrumentation.reservation.request.Device.Emulator.Emulator27
-import com.avito.instrumentation.reservation.request.Device.LocalEmulator
 import com.avito.kotlin.dsl.getOptionalStringProperty
-import com.avito.utils.gradle.KubernetesCredentials
-import com.avito.utils.gradle.kubernetesCredentials
 
 plugins {
     id("com.android.application")
@@ -112,152 +104,3 @@ dependencies(delegateClosureOf<DependencyHandler> {
     androidTestImplementation(Dependencies.test.truth)
     androidTestImplementation(Dependencies.test.okhttpMockWebServer)
 })
-
-extensions.getByType<GradleInstrumentationPluginConfiguration>().apply {
-
-    //todo make these params optional features in plugin
-    reportApiUrl = project.getOptionalStringProperty("avito.report.url") ?: "http://stub"
-    reportApiFallbackUrl = project.getOptionalStringProperty("avito.report.fallbackUrl") ?: "http://stub"
-    reportViewerUrl = project.getOptionalStringProperty("avito.report.viewerUrl") ?: "http://stub"
-    registry = project.getOptionalStringProperty("avito.registry", "registry") ?: "registry"
-    sentryDsn =
-        project.getOptionalStringProperty("avito.instrumentaion.sentry.dsn") ?: "http://stub-project@stub-host/0"
-    slackToken = project.getOptionalStringProperty("avito.slack.token") ?: "stub"
-    fileStorageUrl = project.getOptionalStringProperty("avito.fileStorage.url") ?: "http://stub"
-    // deprecated since 2020.3.6
-    output = project.rootProject.file("outputs/${project.name}/instrumentation").path
-
-    logcatTags = setOf(
-        "UITestRunner:*",
-        "ActivityManager:*",
-        "ReportTestListener:*",
-        "StorageJsonTransport:*",
-        "TestReport:*",
-        "VideoCaptureListener:*",
-        "TestRunner:*",
-        "SystemDialogsManager:*",
-        "AndroidJUnitRunner:*",
-        "ito.android.de:*", //по этому тэгу система пишет логи об использовании hidden/restricted api https://developer.android.com/distribute/best-practices/develop/restrictions-non-sdk-interfaces
-        "*:E"
-    )
-
-    instrumentationParams = mapOf(
-        "videoRecording" to "failed",
-        "jobSlug" to "FunctionalTests"
-    )
-
-
-    val runAllFilterName = "runAll"
-    filters.register(runAllFilterName)
-
-    filters.register("dynamicFilter") {
-        val includeAnnotation: String? = project.getOptionalStringProperty("includeAnnotation", true)
-        if (includeAnnotation != null) {
-            fromSource.includeByAnnotations(setOf(includeAnnotation))
-        }
-        val includePrefix: String? = project.getOptionalStringProperty("includePrefix", true)
-        if (includePrefix != null) {
-            fromSource.includeByPrefixes(setOf(includePrefix))
-        }
-    }
-
-    val defaultFilter = "default"
-    val customFilter: String = project.getOptionalStringProperty("localFilter", defaultFilter)
-
-    configurationsContainer.register("Local") {
-        tryToReRunOnTargetBranch = false
-        reportSkippedTests = true
-        reportFlakyTests = false
-        filter = customFilter
-
-        targetsContainer.register("api27") {
-            deviceName = "API27"
-
-            scheduling = SchedulingConfiguration().apply {
-                quota = QuotaConfiguration().apply {
-                    retryCount = 1
-                    minimumSuccessCount = 1
-                }
-
-                reservation = TestsBasedDevicesReservationConfiguration().apply {
-                    device = LocalEmulator.device(27)
-                    maximum = 1
-                    minimum = 1
-                    testsPerEmulator = 1
-                }
-            }
-        }
-    }
-
-    val credentials = project.kubernetesCredentials
-    if (credentials is KubernetesCredentials.Service || credentials is KubernetesCredentials.Config) {
-        configurationsContainer.register("ui") {
-            tryToReRunOnTargetBranch = false
-            reportSkippedTests = true
-            reportFlakyTests = true
-            filter = customFilter
-
-            targetsContainer.register("api22") {
-                deviceName = "API22"
-
-                scheduling = SchedulingConfiguration().apply {
-                    quota = QuotaConfiguration().apply {
-                        retryCount = 1
-                        minimumSuccessCount = 1
-                    }
-
-                    reservation = TestsBasedDevicesReservationConfiguration().apply {
-                        device = Emulator22
-                        maximum = 50
-                        minimum = 2
-                        testsPerEmulator = 3
-                    }
-                }
-            }
-
-            targetsContainer.register("api27") {
-                deviceName = "API27"
-
-                scheduling = SchedulingConfiguration().apply {
-                    quota = QuotaConfiguration().apply {
-                        retryCount = 1
-                        minimumSuccessCount = 1
-                    }
-
-                    reservation = TestsBasedDevicesReservationConfiguration().apply {
-                        device = Emulator27
-                        maximum = 50
-                        minimum = 2
-                        testsPerEmulator = 3
-                    }
-                }
-            }
-        }
-
-        configurationsContainer.register("uiDebug") {
-            tryToReRunOnTargetBranch = false
-            reportSkippedTests = false
-            reportFlakyTests = false
-            enableDeviceDebug = true
-            filter = customFilter
-
-            targetsContainer.register("api27") {
-                deviceName = "API27"
-
-                scheduling = SchedulingConfiguration().apply {
-                    quota = QuotaConfiguration().apply {
-                        retryCount = 1
-                        minimumSuccessCount = 1
-                    }
-
-                    reservation = TestsBasedDevicesReservationConfiguration().apply {
-                        device = Emulator27
-                        maximum = 1
-                        minimum = 1
-                        testsPerEmulator = 1
-                    }
-                }
-            }
-        }
-    }
-}
